pyproject.toml
[project]
name = "What_Transformer_Looked_At"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Mizuki", email = "mbaba@kanolab.net" }
]
dependencies = [
    "tqdm>=4.66.1",
    "wandb>=0.16.1",
    "pyyaml>=6.0.1",
    "fugashi>=1.3.0",
    "ipadic>=1.0.0",
    "pytest>=7.4.3",
    "torch==1.13.1",
    "pyline-notify @ git+https://github.com/lovelovetrb/pyline_notify",
    "python-dotenv>=1.0.0",
    "sentencepiece>=0.1.99",
    "einops>=0.7.0",
    "matplotlib>=3.7.4",
    "pandas>=2.0.3",
    "transformers @ file:///disk/ssd14tc/mbaba/dev/python/what_transformer_looked_at/src/what_transformer_looked_at/transformers",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/what_transformer_looked_at"]

